requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
choreographer==1.2.1
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
duckdb==1.4.2
fonttools==4.61.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.2.0
kiwisolver==1.4.9
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.23
langchain-openai==0.0.2
langsmith==0.0.87
logistro==2.0.1
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.13.0
numpy==1.26.4
openai==1.109.1
orjson==3.11.5
packaging==23.2
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
rpds-py==0.30.0
seaborn==0.13.2
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.52.1
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
